Move exterior and length to ops
