+++ geopolars/src/geoseries.rs
@@ -1,4 +1,5 @@
 use crate::error::{inner_type_name, GeopolarsError, Result};
+use crate::ops::length::GeodesicLengthMethod;
 #[cfg(feature = "proj")]
 use crate::proj::ProjOptions;
 use crate::util::iter_geom;
@@ -14,12 +15,6 @@ use polars::prelude::{PolarsError, Series};
 use std::convert::Into;
 
 pub type ArrayRef = Box<dyn Array>;
-
-pub enum GeodesicLengthMethod {
-    Haversine,
-    Geodesic,
-    Vincenty,
-}
 
 /// Used to express the origin for a given transform. Can be specified either be with reference to
 /// the geometry being transformed (Centroid, Center) or some arbitrary point.
@@ -295,37 +290,7 @@ impl GeoSeries for Series {
     }
 
     fn euclidean_length(&self) -> Result<Series> {
-        use geo::algorithm::euclidean_length::EuclideanLength;
-        let mut result = MutablePrimitiveArray::<f64>::with_capacity(self.len());
-
-        for geom in iter_geom(self) {
-            let length: f64 = match geom {
-                Geometry::Point(_) => Ok(0.0),
-                Geometry::Line(line) => Ok(line.euclidean_length()),
-                Geometry::LineString(line_string) => Ok(line_string.euclidean_length()),
-                Geometry::Polygon(polygon) => Ok(polygon.exterior().euclidean_length()),
-                Geometry::MultiPoint(_) => Ok(0.0),
-                Geometry::MultiLineString(multi_line_string) => {
-                    Ok(multi_line_string.euclidean_length())
-                }
-                Geometry::MultiPolygon(mutli_polygon) => Ok(mutli_polygon
-                    .iter()
-                    .map(|poly| poly.exterior().euclidean_length())
-                    .sum()),
-                Geometry::GeometryCollection(_) => Err(PolarsError::ComputeError(ErrString::from(
-                    "Length methods are not implemented for geometry collection",
-                ))),
-                Geometry::Rect(rec) => Ok(rec.to_polygon().exterior().euclidean_length()),
-                Geometry::Triangle(triangle) => {
-                    Ok(triangle.to_polygon().exterior().euclidean_length())
-                }
-            }?;
-            result.push(Some(length));
-        }
-
-        let result: PrimitiveArray<f64> = result.into();
-        let series = Series::try_from(("geometry", Box::new(result) as ArrayRef))?;
-        Ok(series)
+        crate::ops::length::euclidean_length(self)
     }
 
     fn explode(&self) -> Result<Series> {
@@ -385,23 +350,7 @@ impl GeoSeries for Series {
     }
 
     fn exterior(&self) -> Result<Series> {
-        let mut output_array = MutableBinaryArray::<i32>::with_capacity(self.len());
-
-        for geom in iter_geom(self) {
-            let maybe_exterior = match geom {
-                Geometry::Polygon(polygon) => {
-                    let exterior: Geometry<f64> = polygon.exterior().clone().into();
-                    Some(exterior.to_wkb(CoordDimensions::xy()).unwrap())
-                }
-                _ => None,
-            };
-            output_array.push(maybe_exterior);
-        }
-
-        let result: BinaryArray<i32> = output_array.into();
-
-        let series = Series::try_from(("geometry", Box::new(result) as ArrayRef))?;
-        Ok(series)
+        crate::ops::exterior::exterior(self)
     }
 
     fn from_geom_vec(geoms: &[Geometry<f64>]) -> Result<Self> {
@@ -422,116 +371,7 @@ impl GeoSeries for Series {
     }
 
     fn geodesic_length(&self, method: GeodesicLengthMethod) -> Result<Series> {
-        use geo::algorithm::{
-            geodesic_length::GeodesicLength, haversine_length::HaversineLength,
-            vincenty_length::VincentyLength,
-        };
-        let mut result = MutablePrimitiveArray::<f64>::with_capacity(self.len());
-
-        let map_vincenty_error =
-            |_| PolarsError::ComputeError(ErrString::from("Failed to calculate vincenty length"));
-
-        for geom in iter_geom(self) {
-            let length: f64 = match (&method, geom) {
-                (_, Geometry::Point(_)) => Ok(0.0),
-
-                (GeodesicLengthMethod::Haversine, Geometry::Line(line)) => {
-                    Ok(line.haversine_length())
-                }
-                (GeodesicLengthMethod::Geodesic, Geometry::Line(line)) => {
-                    Ok(line.geodesic_length())
-                }
-                (GeodesicLengthMethod::Vincenty, Geometry::Line(line)) => {
-                    line.vincenty_length().map_err(map_vincenty_error)
-                }
-
-                (GeodesicLengthMethod::Haversine, Geometry::LineString(line_string)) => {
-                    Ok(line_string.haversine_length())
-                }
-                (GeodesicLengthMethod::Geodesic, Geometry::LineString(line_string)) => {
-                    Ok(line_string.geodesic_length())
-                }
-                (GeodesicLengthMethod::Vincenty, Geometry::LineString(line_string)) => {
-                    line_string.vincenty_length().map_err(map_vincenty_error)
-                }
-
-                (GeodesicLengthMethod::Haversine, Geometry::Polygon(polygon)) => {
-                    Ok(polygon.exterior().haversine_length())
-                }
-                (GeodesicLengthMethod::Geodesic, Geometry::Polygon(polygon)) => {
-                    Ok(polygon.exterior().geodesic_length())
-                }
-                (GeodesicLengthMethod::Vincenty, Geometry::Polygon(polygon)) => polygon
-                    .exterior()
-                    .vincenty_length()
-                    .map_err(map_vincenty_error),
-
-                (_, Geometry::MultiPoint(_)) => Ok(0.0),
-
-                (GeodesicLengthMethod::Haversine, Geometry::MultiLineString(multi_line_string)) => {
-                    Ok(multi_line_string.haversine_length())
-                }
-
-                (GeodesicLengthMethod::Geodesic, Geometry::MultiLineString(multi_line_string)) => {
-                    Ok(multi_line_string.geodesic_length())
-                }
-                (GeodesicLengthMethod::Vincenty, Geometry::MultiLineString(multi_line_string)) => {
-                    multi_line_string
-                        .vincenty_length()
-                        .map_err(map_vincenty_error)
-                }
-                (GeodesicLengthMethod::Haversine, Geometry::MultiPolygon(mutli_polygon)) => {
-                    Ok(mutli_polygon
-                        .iter()
-                        .map(|poly| poly.exterior().haversine_length())
-                        .sum())
-                }
-                (GeodesicLengthMethod::Geodesic, Geometry::MultiPolygon(mutli_polygon)) => {
-                    Ok(mutli_polygon
-                        .iter()
-                        .map(|poly| poly.exterior().geodesic_length())
-                        .sum())
-                }
-
-                (GeodesicLengthMethod::Vincenty, Geometry::MultiPolygon(mutli_polygon)) => {
-                    let result: std::result::Result<Vec<f64>, _> = mutli_polygon
-                        .iter()
-                        .map(|poly| poly.exterior().vincenty_length())
-                        .collect();
-                    result.map(|v| v.iter().sum()).map_err(map_vincenty_error)
-                }
-                (_, Geometry::GeometryCollection(_)) => Err(PolarsError::ComputeError(
-                    ErrString::from("Length methods are not implemented for geometry collection"),
-                )),
-                (GeodesicLengthMethod::Haversine, Geometry::Rect(rec)) => {
-                    Ok(rec.to_polygon().exterior().haversine_length())
-                }
-                (GeodesicLengthMethod::Geodesic, Geometry::Rect(rec)) => {
-                    Ok(rec.to_polygon().exterior().geodesic_length())
-                }
-                (GeodesicLengthMethod::Vincenty, Geometry::Rect(rec)) => rec
-                    .to_polygon()
-                    .exterior()
-                    .vincenty_length()
-                    .map_err(map_vincenty_error),
-                (GeodesicLengthMethod::Haversine, Geometry::Triangle(triangle)) => {
-                    Ok(triangle.to_polygon().exterior().haversine_length())
-                }
-                (GeodesicLengthMethod::Geodesic, Geometry::Triangle(triangle)) => {
-                    Ok(triangle.to_polygon().exterior().geodesic_length())
-                }
-                (GeodesicLengthMethod::Vincenty, Geometry::Triangle(triangle)) => triangle
-                    .to_polygon()
-                    .exterior()
-                    .vincenty_length()
-                    .map_err(map_vincenty_error),
-            }?;
-            result.push(Some(length));
-        }
-
-        let result: PrimitiveArray<f64> = result.into();
-        let series = Series::try_from(("result", Box::new(result) as ArrayRef))?;
-        Ok(series)
+        crate::ops::length::geodesic_length(self, method)
     }
 
     fn geom_type(&self) -> Result<Series> {
